spotted-api/requirements.txt
appdirs==1.4.4
argcomplete==2.0.0
Beaker==1.10.0
beautifulsoup4==4.11.0
blivet==3.4.4
blivet-gui==2.3.0
Brlapi==0.8.4
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
cssselect==0.9.2
cupshelpers==1.0
dasbus==1.6
dbus-python==1.2.18
Deprecated==1.2.13
distro==1.6.0
fedora-third-party==0.10
firebase-admin==5.3.0
Flask==2.2.2
Flask-Cors==3.0.10
fros==1.1
gcloud==0.18.3
google-api-core==2.10.1
google-api-python-client==2.61.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.1
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
gpg==1.17.0
grpcio==1.48.1
grpcio-status==1.48.1
httplib2==0.20.4
humanize==3.13.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.4.2
langtable==0.0.59
libcomps==0.1.18
libvirt-python==8.0.0
lxml==4.7.1
Mako==1.1.4
MarkupSafe==2.1.1
msgpack==1.0.4
nftables==0.1
numpy==1.22.0
oauth2client==4.1.3
olefile==0.46
packaging==21.3
Paste==3.5.0
pexpect==4.8.0
pid==2.2.3
Pillow==9.1.0
ply==3.11
productmd==1.33
proto-plus==1.22.1
protobuf==4.21.6
ptyprocess==0.6.0
pwquality==1.4.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.21.0
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.15.0
pycups==2.0.1
pycurl==7.45.1
pyenchant==3.2.2
PyGObject==3.42.1
pykickstart==3.36
pyOpenSSL==21.0.0
pyparsing==3.0.9
pyparted==3.12.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.1
python-decouple==3.6
python-jwt==3.3.4
python-meh==0.50
pyudev==0.22.0
pyxdg==0.27
regex==2022.9.13
requests==2.28.1
requests-file==1.5.1
requests-ftp==0.3.1
requests-toolbelt==0.9.1
rpm==4.17.1
rsa==4.9
scour==0.38.1
selinux==3.3
sepolicy==3.3
setools==4.4.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sos==4.3
soupsieve==2.3.1
systemd-python==234
Tempita==0.5.2
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
